requirements.txt
transformers>=4.41.1
torch>=2.1.0
Pillow

